move cmd/client/client to rpcclient, rename common/rpcclient to wsclient
