+++ common/rpctypes/utils.go
@@ -1,4 +1,4 @@
-package rpcclient
+package rpctypes
 
 import (
 	"bytes"
@@ -9,8 +9,6 @@ import (
 	"net"
 	"net/http"
 	"time"
-
-	"github.com/noot/atomic-swap/common/rpctypes"
 )
 
 var (
@@ -31,7 +29,7 @@ var (
 )
 
 // PostRPC posts a JSON-RPC call to the given endpoint.
-func PostRPC(endpoint, method, params string) (*rpctypes.Response, error) {
+func PostRPC(endpoint, method, params string) (*Response, error) {
 	data := []byte(`{"jsonrpc":"2.0","method":"` + method + `","params":` + params + `,"id":0}`)
 	buf := &bytes.Buffer{}
 	_, err := buf.Write(data)
@@ -63,7 +61,7 @@ func PostRPC(endpoint, method, params string) (*rpctypes.Response, error) {
 		return nil, fmt.Errorf("failed to read response body: %w", err)
 	}
 
-	var sv *rpctypes.Response
+	var sv *Response
 	if err = json.Unmarshal(body, &sv); err != nil {
 		return nil, err
 	}
