+++ monero/daemon.go
@@ -3,7 +3,7 @@ package monero
 import (
 	"encoding/json"
 
-	"github.com/noot/atomic-swap/common/rpcclient"
+	"github.com/noot/atomic-swap/common/rpctypes"
 )
 
 // DaemonClient represents a monerod client.
@@ -40,7 +40,7 @@ func (c *client) callGenerateBlocks(address string, amount uint) error {
 		return err
 	}
 
-	resp, err := rpcclient.PostRPC(c.endpoint, method, string(params))
+	resp, err := rpctypes.PostRPC(c.endpoint, method, string(params))
 	if err != nil {
 		return err
 	}
