+++ rpcclient/addresses.go
@@ -1,9 +1,9 @@
-package client
+package rpcclient
 
 import (
 	"encoding/json"
 
-	"github.com/noot/atomic-swap/common/rpcclient"
+	"github.com/noot/atomic-swap/common/rpctypes"
 	"github.com/noot/atomic-swap/rpc"
 )
 
@@ -13,7 +13,7 @@ func (c *Client) Addresses() ([]string, error) {
 		method = "net_addresses"
 	)
 
-	resp, err := rpcclient.PostRPC(c.endpoint, method, "{}")
+	resp, err := rpctypes.PostRPC(c.endpoint, method, "{}")
 	if err != nil {
 		return nil, err
 	}
