+++ rpcclient/client.go
@@ -1,6 +1,6 @@
-package client
+package rpcclient
 
-// Client represents a swap client, used to interact with a swap daemon via JSON-RPC calls.
+// Client represents a swap RPC client, used to interact with a swap daemon via JSON-RPC calls.
 type Client struct {
 	endpoint string
 }
