+++ rpcclient/personal.go
@@ -1,9 +1,9 @@
-package client
+package rpcclient
 
 import (
 	"encoding/json"
 
-	"github.com/noot/atomic-swap/common/rpcclient"
+	"github.com/noot/atomic-swap/common/rpctypes"
 	"github.com/noot/atomic-swap/rpc"
 )
 
@@ -22,7 +22,7 @@ func (c *Client) SetSwapTimeout(duration uint64) error {
 		return err
 	}
 
-	resp, err := rpcclient.PostRPC(c.endpoint, method, string(params))
+	resp, err := rpctypes.PostRPC(c.endpoint, method, string(params))
 	if err != nil {
 		return err
 	}
