+++ rpcclient/take_offer.go
@@ -1,10 +1,9 @@
-package client
+package rpcclient
 
 import (
 	"encoding/json"
 	"fmt"
 
-	"github.com/noot/atomic-swap/common/rpcclient"
 	"github.com/noot/atomic-swap/common/rpctypes"
 )
 
@@ -25,7 +24,7 @@ func (c *Client) TakeOffer(maddr string, offerID string, providesAmount float64)
 		return 0, err
 	}
 
-	resp, err := rpcclient.PostRPC(c.endpoint, method, string(params))
+	resp, err := rpctypes.PostRPC(c.endpoint, method, string(params))
 	if err != nil {
 		return 0, err
 	}
